Update build-release-ci.yml
Explanation of the Fix
Removed Repetition: The original log showed: Run dotnet publish... dotnet publish.... The command was likely repeating itself because the multi-line structure (|) was confusing the interpreter.

Removed Line Continuations: We removed all backslashes (\) and placed the entire dotnet publish command on a single logical line. This is the safest way to execute the command using PowerShell on the Windows runner when the command is defined using the YAML multi-line pipe (|) syntax.

Signed-off-by: Joshua Thomas <61997760+neogentrics@users.noreply.github.com>